% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pkg_deps.R
\name{check_pkg_deps}
\alias{check_pkg_deps}
\title{check_pkg_deps}
\usage{
check_pkg_deps(pkg)
}
\arguments{
\item{pkg}{\code{character} name of the package}
}
\value{
\verb{data frame} containing info about version, update date and source of the dependencies
of \code{pkg} currently installed
}
\description{
The function reuturns information about currently installed versions
of dependencies of the package used as input
}
\examples{
\dontrun{
if(interactive()){
 check_pkg_deps("ggplot2")
 }
}
}
