% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_pkg_deps.R
\name{compare_pkg_deps}
\alias{compare_pkg_deps}
\title{compare_pkg_deps}
\usage{
compare_pkg_deps(pkg, depsfile)
}
\arguments{
\item{pkg}{PARAM_DESCRIPTION}

\item{depsfile}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
