% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pkginfo.R
\name{get_pkginfo}
\alias{get_pkginfo}
\title{get_pkginfo}
\usage{
get_pkginfo(pkg)
}
\arguments{
\item{pkg}{\code{character} name of the package}
}
\value{
\verb{data frame} containing info about update date and source of the repo
}
\description{
get information about version of an installed package
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
