% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_pkg_deps.R
\name{save_pkg_deps}
\alias{save_pkg_deps}
\title{save_pkg_deps}
\usage{
save_pkg_deps(pkg, depsfile)
}
\arguments{
\item{pkg}{\code{character} name of the package}

\item{depsfile}{\code{character} file used to save package dapendencies}
}
\description{
Save status of package dependencies to file
}
\examples{
\dontrun{
if(interactive()){
 tmpfile <- tempfile(, fileext = ".RData")
 save_pkg_deps("checkdeps", tmpfile)
 }
}
}
